# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.CreateReportRequest:
    properties:
      description:
        type: string
      end_date:
        type: string
      format:
        $ref: '#/definitions/entity.ReportFormat'
      name:
        type: string
      parameters:
        $ref: '#/definitions/entity.ReportParameters'
      start_date:
        type: string
      type:
        $ref: '#/definitions/entity.ReportType'
    required:
    - end_date
    - format
    - name
    - start_date
    - type
    type: object
  entity.CreateReportScheduleRequest:
    properties:
      description:
        type: string
      format:
        $ref: '#/definitions/entity.ReportFormat'
      frequency:
        $ref: '#/definitions/entity.ReportScheduleFrequency'
      name:
        type: string
      parameters:
        $ref: '#/definitions/entity.ReportParameters'
      recipients:
        items:
          type: string
        type: array
      report_type:
        $ref: '#/definitions/entity.ReportType'
    required:
    - format
    - frequency
    - name
    - recipients
    - report_type
    type: object
...
  entity.ReportFormat:
    enum:
    - CSV
    - EXCEL
    - PDF
    - JSON
    type: string
    x-enum-varnames:
    - ReportFormatCSV
    - ReportFormatExcel
    - ReportFormatPDF
    - ReportFormatJSON
  entity.ReportParameters:
    additionalProperties: true
    type: object
  entity.ReportScheduleFrequency:
    enum:
    - DAILY
    - WEEKLY
    - MONTHLY
    - QUARTERLY
    - YEARLY
    type: string
    x-enum-varnames:
    - ReportScheduleDaily
    - ReportScheduleWeekly
    - ReportScheduleMonthly
    - ReportScheduleQuarterly
    - ReportScheduleYearly
  entity.ReportType:
    enum:
    - INVENTORY
    - PURCHASE
    - SALES
    - PROFIT_LOSS
    - FINANCIAL
    - CUSTOM
    type: string
    x-enum-varnames:
    - ReportTypeInventory
    - ReportTypePurchase
    - ReportTypeSales
    - ReportTypeProfitAndLoss
    - ReportTypeFinancial
    - ReportTypeCustom
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/lugondev/erp-warehouse-simple
  description: Simple ERP Warehouse Management System API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: ERP Warehouse API
  version: "1.0"
paths:
...
  /api/v1/reports:
    get:
      description: List reports based on filter criteria
      parameters:
      - description: Report name
        in: query
        name: name
        type: string
      - description: Report type
        in: query
        name: type
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Report status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List reports
      tags:
      - Reports
    post:
      consumes:
      - application/json
      description: Create a new report
      parameters:
      - description: Report details
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/entity.CreateReportRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new report
      tags:
      - Reports
  /api/v1/reports/{id}:
    delete:
      description: Delete a report by its ID
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a report
      tags:
      - Reports
    get:
      description: Get a report by its ID
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a report by ID
      tags:
      - Reports
  /api/v1/reports/{id}/export:
    post:
      description: Export a report to a specific format (CSV, Excel, PDF)
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: Export format (CSV, EXCEL, PDF)
        in: query
        name: format
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Export a report
      tags:
      - Reports
...
  /api/v1/reports/financial/profit-loss:
    get:
      description: Get profit and loss report
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get profit and loss report
      tags:
      - Reports
  /api/v1/reports/inventory/age:
    get:
      description: Get inventory age report
      parameters:
      - description: Warehouse ID
        in: query
        name: warehouse_id
        type: string
      - description: As of date (YYYY-MM-DD)
        in: query
        name: as_of_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get inventory age report
      tags:
      - Reports
  /api/v1/reports/inventory/value:
    get:
      description: Get inventory value report
      parameters:
      - description: Warehouse ID
        in: query
        name: warehouse_id
        type: string
      - description: As of date (YYYY-MM-DD)
        in: query
        name: as_of_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get inventory value report
      tags:
      - Reports
  /api/v1/reports/purchases/suppliers:
    get:
      description: Get supplier purchase report
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get supplier purchase report
      tags:
      - Reports
  /api/v1/reports/sales/customers:
    get:
      description: Get customer sales report
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get customer sales report
      tags:
      - Reports
  /api/v1/reports/sales/products:
    get:
      description: Get product sales report
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get product sales report
      tags:
      - Reports
  /api/v1/reports/schedules:
    get:
      description: List report schedules based on filter criteria
      parameters:
      - description: Schedule name
        in: query
        name: name
        type: string
      - description: Report type
        in: query
        name: type
        type: string
      - description: Schedule frequency
        in: query
        name: frequency
        type: string
      - description: Active status
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List report schedules
      tags:
      - Reports
    post:
      consumes:
      - application/json
      description: Create a new report schedule
      parameters:
      - description: Report schedule details
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/entity.CreateReportScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new report schedule
      tags:
      - Reports
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"